drm/msm/dp: work around bogus maximum link rate
The OLED display in the Surface Pro 11 reports a maximum link rate of
zero in its DPCD, causing it to fail to probe correctly.

The Surface Pro 11's DSDT table contains some XML with an
"EDPOverrideDPCDCaps" block that defines the max link rate as 0x1E
(8.1Gbps/HBR3).

Hack around this by conditionally overriding the max link rate if its
value is zero.

Signed-off-by: Dale Whinham <daleyo@gmail.com>