drm/edp-panel: Add panel used by Surface Pro 11 (OLED)
The Microsoft Surface Pro 11 (Snapdragon) ships with two panel options;
an LCD model and an OLED model. The OLED model is using the SDC 4195,
aka. Samsung ATNA30DW01-1.

The OLED panel in my SP11 is working fine with panel-edp, however an
error is being logged due to it being unknown.

Unfortunately I don't have access to any documentation for this panel,
so the correct timings are unknown, but I've chosen some values that
seem quite common and appear to be working fine.

Raw EDID for the panel:

00 ff ff ff ff ff ff 00 4c 83 95 41 b0 01 00 06
00 21 01 04 b5 1c 13 78 03 0b d1 af 51 3d b6 23
0b 50 54 00 00 00 01 01 01 01 01 01 01 01 01 01
01 01 01 01 01 01 00 00 00 00 00 00 00 00 00 00
00 00 00 00 00 00 00 00 00 00 00 fd 00 1e 78 ea
ea 48 01 00 00 00 00 00 00 00 00 00 00 00 00 00
00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 fe
00 41 54 4e 41 33 30 44 57 30 31 2d 31 20 01 c4

70 20 79 02 00 20 00 0c ba 41 59 95 41 b0 01 00
06 00 17 00 21 00 1d d6 0a 67 07 40 0b 80 07 00
f2 ea 50 d7 d3 b6 3d 42 0b 01 45 54 b0 60 08 63
18 10 23 78 26 00 09 07 06 03 00 00 00 50 00 00
22 00 28 9a dd 0a 84 3f 0b 9f 00 1f 00 07 00 7f
07 1f 00 07 00 07 00 9a dd 0a 84 3f 0b 9f 00 1f
00 07 00 7f 07 bf 07 07 00 07 00 81 00 0b e3 05
80 00 e6 06 05 01 86 73 01 00 00 00 00 00 07 90

Signed-off-by: Dale Whinham <daleyo@gmail.com>