drm/msm/dp: Add dpcd link_rate quirk for Surface Pro 11 OLED
The Samsung ATNA30DW01-1 OLED panel in Microsoft Surface Pro 11 (Denali)
reports a max link rate value of 0 in the DPCD register, causing the
panel to fail to probe.

Add a quirk for this panel during DPCD read to set the max link rate
to 8.1Gbps (HBR3), which is the expected value as reported by the
"EDPOverrideDPCDCaps" block found in the DSDT (0x1E).

Signed-off-by: Jérôme de Bretagne <jerome.debretagne@gmail.com>